Complete user registration html

+++ src/main/java/com/apress/ravi/Rest/UserRegistrationRestController.java
@@ -1,7 +1,8 @@
 package com.apress.ravi.Rest;
 
 import java.util.List;
-import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,63 +19,89 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.apress.ravi.Exception.CustomErrorType;
 import com.apress.ravi.dto.UserDTO;
 import com.apress.ravi.repository.UserJpaRepository;
 
 @RestController
 @RequestMapping("/api/user")
 public class UserRegistrationRestController {
-	
+
 	public static final Logger logger = LoggerFactory.getLogger(UserRegistrationRestController.class);
-	
+
 	private UserJpaRepository userJpaRepository;
-	
+
 	@Autowired
 	public void setUserJpaRepository(UserJpaRepository userJpaRepository) {
 		this.userJpaRepository = userJpaRepository;
 	}
-	
+
 	@GetMapping("/")
-	public ResponseEntity<List<UserDTO>>listAllUsers(){
+	public ResponseEntity<List<UserDTO>> listAllUsers() {
 		List<UserDTO> users = userJpaRepository.findAll();
+		if (users.isEmpty()) {
+			return new ResponseEntity<List<UserDTO>>(HttpStatus.NO_CONTENT);
+		}
 		return new ResponseEntity<List<UserDTO>>(users, HttpStatus.OK);
 	}
-	
+
 	@PostMapping(value = "/", consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<UserDTO> createUser(@RequestBody final UserDTO user){
-		
+	public ResponseEntity<UserDTO> createUser(@Valid @RequestBody final UserDTO user) {
+		if (userJpaRepository.findByName(user.getName()) != null) {
+			return new ResponseEntity<UserDTO>(
+					new CustomErrorType(
+							"Unable to create new user.  " + "A User with name " + user.getName() + " already exists."),
+					HttpStatus.CONFLICT);
+		}
+
 		userJpaRepository.save(user);
 		return new ResponseEntity<UserDTO>(user, HttpStatus.CREATED);
-		
+
 	}
-	
+
 	@GetMapping("/{id}")
-	public ResponseEntity<UserDTO> getUserById(@PathVariable("id") final Long id){
+	public ResponseEntity<UserDTO> getUserById(@PathVariable("id") final Long id) {
 		UserDTO user = userJpaRepository.findById(id).orElse(null);
+
+		if (user == null) {
+			return new ResponseEntity<UserDTO>(new CustomErrorType("User with id " + id + " not found"),
+					HttpStatus.NOT_FOUND);
+		}
 		return new ResponseEntity<UserDTO>(user, HttpStatus.OK);
 	}
-	
+
 	@PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<UserDTO> updateUser(@PathVariable("id") final Long id, @RequestBody UserDTO user) {
-		
-		UserDTO currentUser = (UserDTO)userJpaRepository.findAllById(id);
-		
+
+		UserDTO currentUser = (UserDTO) userJpaRepository.findAllById(id);
+
+		if (currentUser == null) {
+			return new ResponseEntity<UserDTO>(
+					new CustomErrorType("Unable to update. User with id " + id + " not found."), HttpStatus.NOT_FOUND);
+		}
+
 		currentUser.setName(user.getName());
 		currentUser.setPassword(user.getPassword());
 		currentUser.setAddress(user.getAddress());
 		currentUser.setEmail(user.getEmail());
-		
+
 		userJpaRepository.saveAndFlush(currentUser);
-		
+
 		return new ResponseEntity<UserDTO>(currentUser, HttpStatus.OK);
 	}
-	
-		
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<UserDTO> deleteUser(@PathVariable("id") final Long id){
+	public ResponseEntity<UserDTO> deleteUser(@PathVariable("id") final Long id) {
+		UserDTO user = userJpaRepository.findAllById(id);
+
+		if (user == null) {
+			return new ResponseEntity<UserDTO>(
+					new CustomErrorType("Unable to delete.  User with id " + id + " not found."), HttpStatus.NOT_FOUND);
+
+		}
+
 		userJpaRepository.deleteById(id);
 		return new ResponseEntity<UserDTO>(HttpStatus.NO_CONTENT);
 	}
-	
 
 }
